requirements.txt
asgiref==3.3.1
Django==3.1.6
django-crispy-forms==1.11.0
django-debug-toolbar==3.2
django-jazzmin==2.4.4
django-prometheus==2.1.0
djangorestframework==3.12.2
Pillow==8.1.0
prometheus-client==0.9.0
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
